requirements_dev.txt
-r requirements.txt
ddt
pytest
pytest-django
pytest-django-model
pytest-django-testing-postgresql
pytest-djangorestframework
